+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - The lock now can be released with async context manager.
 - Support the same address formats as aioredis does.
 - ``Aioredlock.is_locked()`` is implemented to check is the resource acquired by another redlock instance.
+- The ``lock_timeout`` and ``drift`` parameters now mesured in seconds, just like ``retry_delay_min`` and ``retry_delay_max``.
 
 ## 0.1.1 - 2017-05-14